pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page analyzer is a website that analyzes the specified pages for SEO suitability"
authors = ["Konstantin Lugovskikh <kostya.lugovskikh@mail.ru>"]
readme = "README.md"
packages = [
  { include = "page_analyzer"},
]

[tool.poetry.dependencies]
python = "^3.10"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
validators = "^0.22.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
flask = "^3.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
